requirements.txt
flask==3.0.3
openai==1.59.7
python-dotenv==1.0.1
requests==2.32.3
